refactor(login-faker): updated path of login faker files
